LayoutTests/fast/viewport/viewport-72.html

<head>
    <title>Check that property name without a value does not cause synchronization on following semi-colon.</title>
    <meta name="viewport" content="width=100 x;initial-scale=1">
    <script>
        function test() {
            if (window.layoutTestController) {
                layoutTestController.dumpAsText();
                layoutTestController.dumpConfigurationForViewport(320, 352);
            }
        }
    </script>
</head>
<body onload="test();">
